Make the code more modular
Add env checker
