+++ dev-tools/build_release.py
@@ -14,13 +14,10 @@
 # either express or implied. See the License for the specific
 # language governing permissions and limitations under the License.
 
-import re
 import os
 import argparse
-import subprocess
 import sys
 
-from functools import partial
 
 from os.path import dirname, abspath
 from release.logger import logger
@@ -29,6 +26,7 @@ from release.git import git
 from release.announcement import announcement
 from release.utils import run
 from release.utils import checksums
+from release.utils import envcheck
 from release.documentation import doc_update
 from release.maven import mvn
 
@@ -78,12 +76,6 @@ ROOT_DIR = abspath(os.path.join(abspath(dirname(__file__)), '../'))
 README_FILE = ROOT_DIR + '/README.md'
 POM_FILE = ROOT_DIR + '/pom.xml'
 DEV_TOOLS_DIR = ROOT_DIR + '/plugin_tools'
-
-# console colors
-OKGREEN = '\033[92m'
-OKWARN = '\033[93m'
-ENDC = '\033[0m'
-FAIL = '\033[91m'
 
 
 ##########################################################
@@ -162,70 +154,6 @@ def check_s3_credentials():
             'Could not find "AWS_ACCESS_KEY_ID" / "AWS_SECRET_ACCESS_KEY" in the env variables please export in order to upload to S3')
 
 
-def check_command_exists(name, cmd):
-    try:
-        print('%s' % cmd)
-        subprocess.check_output(cmd, shell=True, stderr=subprocess.STDOUT)
-    except subprocess.CalledProcessError:
-        raise RuntimeError('Could not run command %s - please make sure it is installed' % (name))
-
-
-def run_and_print(text, run_function, optional=False):
-    try:
-        print(text, end='')
-        run_function()
-        print(OKGREEN + 'OK' + ENDC)
-    except RuntimeError:
-        if optional:
-            print(OKWARN + 'NOT PRESENT' + ENDC)
-        else:
-            print(FAIL + 'FAILED' + ENDC)
-
-
-def check_env_var(text, env_var, optional=False):
-    try:
-        print(text, end='')
-        env[env_var]
-        print(OKGREEN + 'OK' + ENDC)
-    except KeyError:
-        if optional:
-            print(OKWARN + 'NOT PRESENT' + ENDC)
-        else:
-            print(FAIL + 'FAILED' + ENDC)
-
-
-def check_environment_and_commandline_tools():
-    check_env_var('Checking for AWS env configuration AWS_SECRET_ACCESS_KEY_ID...     ', 'AWS_SECRET_ACCESS_KEY')
-    check_env_var('Checking for AWS env configuration AWS_ACCESS_KEY_ID...            ', 'AWS_ACCESS_KEY_ID')
-    check_env_var('Checking for Github env configuration GITHUB_LOGIN...              ', 'GITHUB_LOGIN', True)
-    check_env_var('Checking for Github env configuration GITHUB_PASSWORD...           ', 'GITHUB_PASSWORD', True)
-    check_env_var('Checking for Github env configuration GITHUB_KEY...                ', 'GITHUB_KEY', True)
-    check_env_var('Checking for Email settings MAIL_SENDER...                         ', 'MAIL_SENDER')
-    check_env_var('Checking for Email settings MAIL_TO...                             ', 'MAIL_TO', True)
-    check_env_var('Checking for Email settings SMTP_SERVER...                         ', 'SMTP_SERVER', True)
-    # check_env_var('Checking for SONATYPE env configuration SONATYPE_USERNAME...       ', 'SONATYPE_USERNAME')
-    # check_env_var('Checking for SONATYPE env configuration SONATYPE_PASSWORD...       ', 'SONATYPE_PASSWORD')
-    # check_env_var('Checking for GPG env configuration GPG_KEY_ID...                   ', 'GPG_KEY_ID')
-    # check_env_var('Checking for GPG env configuration GPG_PASSPHRASE...               ', 'GPG_PASSPHRASE')
-    # check_env_var('Checking for S3 repo upload env configuration S3_BUCKET_SYNC_TO... ', 'S3_BUCKET_SYNC_TO')
-    # check_env_var('Checking for git env configuration GIT_AUTHOR_NAME...              ', 'GIT_AUTHOR_NAME')
-    # check_env_var('Checking for git env configuration GIT_AUTHOR_EMAIL...             ', 'GIT_AUTHOR_EMAIL')
-
-    run_and_print('Checking command: gpg...            ', partial(check_command_exists, 'gpg', 'gpg --version'))
-    run_and_print('Checking command: expect...         ', partial(check_command_exists, 'expect', 'expect -v'))
-    # run_and_print('Checking command: createrepo...     ', partial(check_command_exists, 'createrepo', 'createrepo --version'))
-    run_and_print('Checking command: s3cmd...          ', partial(check_command_exists, 's3cmd', 's3cmd --version'))
-    # run_and_print('Checking command: apt-ftparchive... ', partial(check_command_exists, 'apt-ftparchive', 'apt-ftparchive --version'))
-
-    # boto, check error code being returned
-    location = os.path.dirname(os.path.realpath(__file__))
-    command = 'python %s/upload-s3.py' % location
-    run_and_print('Testing boto python dependency...   ', partial(check_command_exists, 'python-boto', command))
-
-    run_and_print('Checking java version...            ', partial(mvn.verify_java_version, '1.7'))
-    run_and_print('Checking java mvn version...        ', partial(mvn.verify_mvn_java_version, '1.7', mvn.MVN))
-
-
 if __name__ == '__main__':
     parser = argparse.ArgumentParser(description='Builds and publishes a Elasticsearch Plugin Release')
     parser.add_argument('--branch', '-b', metavar='master', default=git.get_current_branch(),
@@ -254,7 +182,7 @@ if __name__ == '__main__':
     mail = args.mail
 
     if args.check:
-        check_environment_and_commandline_tools()
+        envcheck.check_environment_and_commandline_tools()
         sys.exit(0)
 
     if src_branch == 'master':
